refactor: Split util.go into multiple files
