+++ netbox/util_virtualization.go
@@ -1,0 +1,117 @@
+package netbox
+
+import (
+	"fmt"
+	"strconv"
+
+	netboxclient "github.com/smutel/go-netbox/v3/netbox/client"
+	"github.com/smutel/go-netbox/v3/netbox/client/virtualization"
+	"github.com/smutel/go-netbox/v3/netbox/models"
+	"github.com/smutel/terraform-provider-netbox/v4/netbox/internal/requestmodifier"
+)
+
+// Type of vm interface in Netbox
+const vMInterfaceType string = "virtualization.vminterface"
+
+func getVMIDForInterface(m interface{}, objectID int64) (int64, error) {
+	client := m.(*netboxclient.NetBoxAPI)
+
+	objectIDStr := fmt.Sprintf("%d", objectID)
+	paramsInterface := virtualization.NewVirtualizationInterfacesListParams().WithID(
+		&objectIDStr)
+	interfaces, err := client.Virtualization.VirtualizationInterfacesList(
+		paramsInterface, nil)
+
+	if err != nil {
+		return 0, err
+	}
+
+	for _, i := range interfaces.Payload.Results {
+		if i.ID == objectID {
+			if i.VirtualMachine != nil {
+				return i.VirtualMachine.ID, nil
+			}
+		}
+	}
+	return 0, fmt.Errorf("virtual machine not found")
+}
+
+func isprimary(m interface{}, objectID *int64, ipID int64, ip4 bool) (bool, error) {
+	client := m.(*netboxclient.NetBoxAPI)
+
+	if objectID == nil {
+		return false, nil
+	}
+
+	objectIDStr := strconv.FormatInt(*objectID, 10)
+	paramsInterface := virtualization.NewVirtualizationInterfacesListParams().WithID(
+		&objectIDStr)
+	interfaces, err := client.Virtualization.VirtualizationInterfacesList(
+		paramsInterface, nil)
+
+	if err != nil {
+		return false, err
+	}
+
+	for _, i := range interfaces.Payload.Results {
+		if i.ID == *objectID {
+			if i.VirtualMachine != nil {
+				vmIDStr := fmt.Sprintf("%d", i.VirtualMachine.ID)
+				paramsVM := virtualization.NewVirtualizationVirtualMachinesListParams().WithID(&vmIDStr)
+				vms, err := client.Virtualization.VirtualizationVirtualMachinesList(
+					paramsVM, nil)
+
+				if err != nil {
+					return false, err
+				}
+
+				if *vms.Payload.Count != 1 {
+					return false, fmt.Errorf("Cannot set an interface as primary when " +
+						"the interface is not attached to a VM.")
+				}
+
+				if ip4 && vms.Payload.Results[0].PrimaryIp4 != nil {
+					return vms.Payload.Results[0].PrimaryIp4.ID == ipID, nil
+				} else if !ip4 && vms.Payload.Results[0].PrimaryIp6 != nil {
+					return vms.Payload.Results[0].PrimaryIp6.ID == ipID, nil
+				} else {
+					return false, nil
+				}
+			} else {
+				return false, fmt.Errorf("Cannot set an interface as primary when the " +
+					"interface is not attached to a VM.")
+			}
+		}
+	}
+
+	return false, nil
+}
+
+func updatePrimaryStatus(m interface{}, vmid, ipid int64, primary bool) error {
+	client := m.(*netboxclient.NetBoxAPI)
+
+	emptyFields := make(map[string]interface{})
+	dropFields := []string{
+		"created",
+		"last_updated",
+		"name",
+		"cluster",
+		"tags",
+	}
+
+	params := &models.WritableVirtualMachineWithConfigContext{}
+	if primary {
+		params.PrimaryIp4 = &ipid
+	} else {
+		params.PrimaryIp4 = nil
+		emptyFields["primary_ip4"] = nil
+	}
+	vm := virtualization.NewVirtualizationVirtualMachinesPartialUpdateParams().WithData(params)
+	vm.SetID(vmid)
+	_, err := client.Virtualization.VirtualizationVirtualMachinesPartialUpdate(
+		vm, nil, requestmodifier.NewRequestModifierOperation(emptyFields, dropFields))
+	if err != nil {
+		return err
+	}
+	return nil
+}
